==> 10_publication_tables.Rmd <==
---
title: "Tables for Publication"
output: html_notebook
---

# Load Libraries

```{r libraries}
library(flextable) # build stats table
library(tidyverse) # for data wrangling
library(officer) # to format to save as docx
```

# Table 1

**Table 1** Results of linear mixed effect (LME) models (site as a random factor) assessing spatial and temporal variation in vibratory noise amplitude (measured as Leq in units dB re FS).

```{r table 1}
noise_stats <- read.csv(file = "data/vibratory_noise_stats.csv", header = TRUE) # for predictor stats table

noise_stats <- noise_stats %>% 
  mutate(Variable = factor(Variable),
         Variable = fct_recode(Variable, "Rural/Urban" = "Category", "Potential Traffic Impact (PC1)" = "PC1", 
                               "Potential Traffic Impact (PC1) x Substrate" = "PC1xSubstrate",
                               "Visit x Rural/Urban" = "VisitxCategory",
                               "Precent Weekly Harvest" = "PerWeeklyHarvest"),
         Data = factor(Data),
         Data = fct_recode(Data, "Daily Average" = "DailyAvg1",
                                "Daily Average " = "DailyAvg2",
                                "Daily Average (Rural)" = "Rural",
                                "Daily Average (Urban)" = "Urban",
                                "Daily Average (2020)" = "DayAvg2020",
                                "Daily Average (2020, Rural)" = "DayAvg2020Rural"))

noise_stats[noise_stats == 0] <- ""

noise_stats <- noise_stats %>% 
  mutate(Pvalue = factor(Pvalue),
         Pvalue = fct_recode(Pvalue, "< 0.001 ***" = "0.001",
                        "   0.490" = "0.49",
                        "   0.002 **" = "0.002",
                        "   0.567" = "0.567",
                        "   0.935" = "0.935",
                        "   0.061 ." = "0.061",
                        "   0.146" = "0.146",
                        "   0.065 ." = "0.065",
                        "   0.318" = "0.318",
                        "   0.322" = "0.322"))

colnames(noise_stats) <- c('Noise Variation','Data (Subset by)','Variable', 'Test Statistic', 'N', 'P-Value', 'Cond. R²', 'Marg. R²')

sect_properties <- prop_section(
  page_size = page_size(orient = "portrait",
                        width = 8.5, height = 11),
  type = "continuous")

FitFlextableToPage <- function(ft, pgwidth = 6){

  ft_out <- ft %>% autofit()

  ft_out <- width(ft_out, width = dim(ft_out)$widths*pgwidth /(flextable_dim(ft_out)$widths))
  return(ft_out)
}

FitFlextableToPage(flextable(noise_stats[ , 1:8]) %>% 
  autofit() %>% 
  merge_v(j = c('Noise Variation', 'Data (Subset by)', 'Cond. R²', 'Marg. R²'), part = "body") %>% 
  bold(bold = TRUE, part = "header") %>% 
  hline_bottom(part = "body", border = officer::fp_border(width = 2)) %>% 
  border(j = 'Noise Variation', border.bottom = officer::fp_border(width = 2)) %>% 
  hline(i = c("10"), part = "body", border = officer::fp_border(width = 2)) %>%   
  hline(i = c("1", "4", "7", "13"), j = c("2":"8"), part = "body", border = officer::fp_border(width = 1, color = "grey")) %>%     
  fontsize(size = 8, part = "all") %>% 
  align(align = "center", part = "all") %>% 
  valign(valign = "top", part = "all") %>% 
  footnote(i = c(5:7, 14), j = 4, ref_symbols = "a", value = as_paragraph("Results from a robust linear mixed effects model")) %>% 
    footnote(i = c(11:13), j = 4, ref_symbols = "b", value = as_paragraph("Results from a Type II Wald Chi-Squared test of a linear mixed effect model (test statistic = Chi-Squared)"))) %>% 
  save_as_docx(path = here::here("figures/table1.docx"),  pr_section = sect_properties)
```

# Table 2

**Table 2** Final model results of the statistical analyses investigating predictors of spiders' microhabitat use under a variable vibratory environment.

```{r table 2}
choice_stats <- read.csv(file = "data/choice_stats.csv", header = TRUE)

choice_stats <- choice_stats %>% 
  filter(Distibuion != "neg_binom") %>% 
  mutate(Focal = factor(Focal),
         Focal = fct_recode(Focal, "Spider Position" = "Spider", 
                            "Silk Mass" = "Silk"),
         Distribution = factor(Distibuion), 
         Distribution = fct_recode(Distribution, "Binomial" = "binomial",
                                   "Binomial " = "binomial2", 
                                   "Negative Binomial" = "neg_binom",
                                   "Beta" = "beta"),
         Response = factor(Data),
         Response = fct_recode(Response, "(iv) Stay (1) or Move (0)" = "stay",
                               "(i) Loud (1) or Quiet (0)" = "side",
                               "(ii) Loud (1) or Quiet (0)" = "silk_side",
                               "(iii) Proportion of Silk Mass" = "prop"), 
         Predictor = factor(Variable),
         Predictor = fct_recode(Predictor, "Amplitude" = "leq",
                               "Body Condition" = "cond",
                               "Age" = "age",
                               "Chamber" = "part",
                               "Amplitude x Body Condition" = "leqxcond",
                               "Amplitude x Age" = "leqxage",
                               "Chamber x Amp." = "partxleq",
                               "Chamber x Age" = "partxage",
                               "Chamber x Amp. x Age" = "partxleqxage"))

choice_stats <- choice_stats %>% 
  mutate(Pvalue = factor(Pvalue),
         Pvalue = fct_recode(Pvalue, " < 0.001 ***" = "0.001",
                        "   0.460" = "0.46",
                        "   0.910" = "0.91",
                        "   0.120" = "0.12",
                        "   0.761" = "0.761",
                        "   0.718" = "0.718",
                        "   0.081 ." = "0.081",
                        "   0.268" = "0.268",
                        "   0.469" = "0.469",
                        "   0.053 ." = "0.053",
                        "   0.110" = "0.11",
                        "   0.306" = "0.306",
                        "   0.843" = "0.843",
                        "   0.044 *" = "0.044",
                        "   0.818" = "0.818",
                        "   0.009 **" = "0.009")) %>% 
  dplyr::select(Focal, Distribution, Response, Predictor, z, resid.df, Pvalue, R2)

colnames(choice_stats) <- c(' ','Distribution','Response', 'Predictor', 'Test Statistic', 'Residual DF', 'P-Value', 'R²')

FitFlextableToPage(flextable(choice_stats[, 1:8]) %>% 
  autofit() %>% 
  merge_v(j = c(' ', 'Distribution', 'Response', 
                'Test Statistic', 'Residual DF', 
                'R²'), part = "body") %>% 
  theme_vanilla() %>% 
  hline(i = c("1", "2", "4", "5", "7", "8", "9", "10", "11", "13", "14"), 
        j = c("Predictor", "Test Statistic", "P-Value"), 
        part = "body", 
        border = officer::fp_border(width = 0, color = "black")) %>% 
  bold(bold = TRUE, part = "header") %>% 
  fontsize(size = 8, part = "all") %>% 
  align(align = "left", part = "all") %>%
  #align(j = "P-Value", align = "left", part = "body") %>% 
  valign(valign = "top", part = "all") %>% 
    footnote(i = c(1), j = 5, part = "header", ref_symbols = "a", value = as_paragraph("The test statistic was a z-value for all tests except for beta regression where we provide a chi-square value.")) %>%
  footnote(i = c(1), j = 8, ref_symbols = "b", value = as_paragraph("This test was a mixed model with spider ID as a random effect. As such, the R² value is the conditional R². The marginal R² value was 0.055."))) %>% 
  save_as_docx(path = here::here("figures/table2.docx"),  pr_section = sect_properties)
```
